Convert withHotkeys to hooks
Part 1
