+++ plugiamo/src/app/content/content-frame.js
@@ -1,6 +1,6 @@
 import CloseButton from './close-button'
 import styled from 'styled-components'
-import withHotkeys, { escapeKey } from 'ext/recompose/with-hotkeys'
+import withHotkeys, { escapeKey } from 'ext/hooks/with-hotkeys'
 import { compose } from 'recompose'
 import { Frame, positioning, useAnimateOnMount } from 'plugin-base'
 import { h } from 'preact'
