+++ plugiamo/src/app/index.js
@@ -9,7 +9,7 @@ import mixpanel from 'ext/mixpanel'
 import Router from './content/router'
 import setup, { optionsFromHash } from './setup'
 import setupFlowHistory from './setup/flow-history'
-import withHotkeys, { escapeKey } from 'ext/recompose/with-hotkeys'
+import withHotkeys, { escapeKey } from 'ext/hooks/with-hotkeys'
 import { assessmentCart, assessmentHack } from 'special/assessment/utils'
 import { compose, lifecycle, withHandlers, withProps, withState } from 'recompose'
 import { gql, graphql } from 'ext/recompose/graphql'
