+++ plugiamo/src/special/assessment/size-guide/index.js
@@ -8,7 +8,7 @@ import googleAnalytics from 'ext/google-analytics'
 import Launcher from 'app/launcher'
 import mixpanel from 'ext/mixpanel'
 import useChatActions from 'ext/hooks/use-chat-actions'
-import withHotkeys, { escapeKey } from 'ext/recompose/with-hotkeys'
+import withHotkeys, { escapeKey } from 'ext/hooks/with-hotkeys'
 import { branch, compose, lifecycle, renderNothing, withHandlers, withProps, withState } from 'recompose'
 import { fetchProducts } from 'special/assessment/utils'
 import { getScrollbarWidth, isSmall } from 'utils'
