+++ plugiamo/src/special/spotahome/index.js
@@ -2,7 +2,7 @@ import AppBase from 'app/base'
 import Base from './base'
 import Launcher from 'app/launcher'
 import mixpanel from 'ext/mixpanel'
-import withHotkeys, { escapeKey } from 'ext/recompose/with-hotkeys'
+import withHotkeys, { escapeKey } from 'ext/hooks/with-hotkeys'
 import { branch, compose, lifecycle, renderNothing, withHandlers, withProps, withState } from 'recompose'
 import { gql } from 'ext/recompose/graphql'
 import { h } from 'preact'
